Move config, predicted_stats and sleeper_api into the package
These three were sitting at the repo root looking like leftovers from the
refactor, but they are live code: 19 import statements across 11 files
depend on them. So this is a move plus an import rewrite, not a deletion.

They are now refactored/config.py, refactored/predicted_stats.py and
refactored/sleeper_api.py, imported relatively from inside the package and as
refactored.<mod> from tests. Checked first that neither DATA_DIR nor
load_dotenv() resolves relative to __file__ -- DATA_DIR is "./data", relative
to the working directory -- so the cache still lands in the same place and the
Docker volume mount is unaffected.

Four tests failed on the first pass because @patch() targets are strings and
don't move with the module; "sleeper_api.get_league" is now
"refactored.sleeper_api.get_league".

Also collapses the sys.path bootstrap that had been copy-pasted into the top
of all six test modules into a single tests/conftest.py. pytest loads conftest
before collection, so imports in the test files now sit at the top of the file
where they belong, and the E402 exemption they needed is gone from pyproject.

Verified: ruff check + format clean, 58/58 tests pass both as
`python -m pytest tests/` from the root and as `pytest` from inside tests/,
and GET /health on the WSGI app still returns 200.
