+++ refactored/predicted_stats.py
