+++ tests/test_api.py
@@ -1,14 +1,6 @@
 import json
-import os
-import sys
 import unittest
 from unittest.mock import patch
-
-# Ensure project root is on sys.path so 'refactored' can be imported when tests
-# are executed from the tests/ directory or other working dirs.
-PROJECT_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-if PROJECT_ROOT not in sys.path:
-    sys.path.insert(0, PROJECT_ROOT)
 
 from refactored.api import application
