+++ tests/test_draft_prep.py
@@ -1,12 +1,6 @@
 import datetime as dt
-import os
-import sys
 import unittest
 from unittest.mock import patch
-
-PROJECT_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-if PROJECT_ROOT not in sys.path:
-    sys.path.insert(0, PROJECT_ROOT)
 
 from refactored import draft_prep
