+++ tests/test_range_model.py
@@ -1,10 +1,4 @@
-import os
-import sys
 import unittest
-
-PROJECT_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-if PROJECT_ROOT not in sys.path:
-    sys.path.insert(0, PROJECT_ROOT)
 
 from refactored import range_model
 from refactored.prob_models import poisson_quantile
