+++ tests/test_weekly_windows.py
@@ -1,11 +1,5 @@
 import datetime as dt
-import os
-import sys
 import unittest
-
-PROJECT_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-if PROJECT_ROOT not in sys.path:
-    sys.path.insert(0, PROJECT_ROOT)
 
 from refactored import weekly_windows as ww
